+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.5 (07-Apr-2015)
+
+  * Add version number to executable
+  * Add .rc and .ico files to build
+  * Added copyright information for jpgd093b library to LICENSE
+  * Added jpgd093b library source code
+
 ## 1.4 (27-Mar-2015)
 
   * Executables signed with StartSSL code signing certificate